+++ src/postulo/accounts/views.py
@@ -67,6 +67,7 @@ class ProfileView(LoginRequiredMixin, UpdateView):
         kwargs = {
             "holder": self.object,
             "default_country": phones.default_country(getattr(self.object, "language", "")),
+            "asked_by": self.request.user,
         }
         if self.request.method == "POST" and "phone_numbers-TOTAL_FORMS" in self.request.POST:
             kwargs["data"] = self.request.POST